add firewalld command to vault installation

+++ bin/prepare.sh
@@ -5,16 +5,14 @@
 # as an OVA
 #
 
-OVFCONF_RAN="/etc/ovfconf"
-SSH_KNOWN_HOSTS_FILE="/root/.ssh/known_hosts"
-BASH_HISTORY_FILE="/root/.bash_history"
-GITCONFIG_FILE="/root/.gitconfig"
-
 #
 # VMware KB 82228
 #
-echo "Resetting the machine ID"
-echo -n > /etc/machine-id
+if [ -f  /etc/machine-id ]; then
+  echo "Resetting the machine ID"
+  echo -n > /etc/machine-id
+fi
+
 if [ -f /var/lib/dbus/machine-id ]; then
   echo "Reset the dbus machine-id"
   rm -f /var/lib/dbus/machine-id
@@ -22,18 +20,21 @@ if [ -f /var/lib/dbus/machine-id ]; then
 fi
 
 # Bash history
+BASH_HISTORY_FILE="/root/.bash_history"
 if [ -f ${BASH_HISTORY_FILE} ]; then
   echo "Deleting bash history"
   rm -f "${BASH_HISTORY_FILE}"
 fi
 
 # SSH known hosts
+SSH_KNOWN_HOSTS_FILE="/root/.ssh/known_hosts"
 if [ -f ${SSH_KNOWN_HOSTS_FILE} ]; then
   echo "Deleting known hosts"
   rm -f "${SSH_KNOWN_HOSTS_FILE}"
 fi
 
 # Gitconfig
+GITCONFIG_FILE="/root/.gitconfig"
 if [ -f ${GITCONFIG_FILE} ]; then
   echo "Deleting gitconfig"
   rm -f "${GITCONFIG_FILE}"
@@ -60,6 +61,7 @@ find /var/log -type f -regex ".*\.gz$" -delete
 find /var/log -type f -regex ".*\.[0-9]$" -delete
 
 # Remove ovfconf
+OVFCONF_RAN="/etc/ovfconf"
 echo "Removing ovfconf file"
 rm -f "${OVFCONF_RAN}"
